Initial test on Html and Text parser.
Some initial refactoring on both classes in order to return an array of sessions which contains various information such as the date, participants and the list of messages that were sent. Each message entry should contain the time the message was sent, the nick of the sender (may be complete or partial depending on the input), the email of the sender and the message itself.